fuzzerbot/pyproject.toml
[tool.poetry]
name = "fuzzerbot"
version = "0.1.0"
description = ""
authors = ["Martin Hloska <martin.hloska@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
"discord.py" = "^2.4.0"
watchdog = "^6.0.0"
python-dotenv = "^1.0.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
